+++ processor/src/test/resources/bad_input/com/example/computed_value_ref/FinalModel.java
@@ -1,0 +1,23 @@
+package com.example.computed_value_ref;
+
+import javax.annotation.Nonnull;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.Computed;
+import org.realityforge.arez.annotations.ComputedValueRef;
+
+@ArezComponent
+public class FinalModel
+{
+  @Computed
+  public long getTime()
+  {
+    return 0;
+  }
+
+  @Nonnull
+  @ComputedValueRef
+  final org.realityforge.arez.ComputedValue getTimeComputedValue()
+  {
+    throw new IllegalStateException();
+  }
+}