+++ processor/src/test/resources/bad_input/com/example/computed_value_ref/NoComputedValueModel.java
@@ -1,0 +1,16 @@
+package com.example.computed_value_ref;
+
+import javax.annotation.Nonnull;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.ComputedValueRef;
+
+@ArezComponent
+public class NoComputedValueModel
+{
+  @Nonnull
+  @ComputedValueRef
+  public org.realityforge.arez.ComputedValue getTimeComputedValue()
+  {
+    throw new IllegalStateException();
+  }
+}